Listing Info
General layout for the listing information has been implemented.  There
are some tweaks that will need to be fixed with the search bar up top
so that it shrinks when going into mobile size window.  The drop down
for Home Type is also not working correctly with the mobile version.