Listing Page mobile update
Mobile has been formatted to a slightly more stable layout.

Things To Fix:
1) Rows for the grid list are not appearing as desired.

2) Changed font sizes for screen sizes.  This kind of helps headers and
chart info but needs more fixing to actually be “responsive.”

3) Noticed that the header image is not really displayed all that great
at the top of the page in tablet and mobile screen sizes.  Look into
revising.

4) Look into fixing the Navbar as the screen sizes change.  Mobile has
the worst affect of navbar as of now.  Search bar must shrink and the
font color needs to change to white as it is hardly visible in the
dropdown menus.