Area Page Update
Areas page has been updated so that it will be displayed similarly to
the Listing page.  It is also responsive in the same way as the Listing
page for tablet and mobile device sizes.

Things To Fix:
1) Rows for the grid list are not appearing as desired.

2) Noticed that the header image is not really displayed all that great
at the top of the page in tablet and mobile screen sizes.  Look into
revising.

3) Look into fixing the Navbar as the screen sizes change.  Mobile has
the worst affect of navbar as of now.  Search bar must shrink and the
font color needs to change to white as it is hardly visible in the
dropdown menus.