Started Organizing
No new code, but there are new files and directories to help organize
the pages and files better.  Creating ejs files and trying to get the
ejs pages implemented using templates.  All of the css files need to be
revised so that code is not repeated and some css code could be put
into its own file for easier organization.