Cleaning ListingGrid
Listing info on the grid has been tweaked to flow with the photos.
Colors are not set in stone yet and a link to the listing page will
need to be added to each photo. The app.css file will need to be
cleaned up as it has a lot of dead code.

**Still need to fix the status label.